Fold actual layer into exports
