--- packages/bundle/src/boot/exports/index.ts
@@ -1,5 +1,0 @@
-// Default is not re-exported when using `export *`.
-import ReactWebChat from '../actual/full';
-
-export * from '../actual/full';
-export default ReactWebChat;
