+++ packages/bundle/src/boot/exports/middleware.ts
@@ -1,1 +1,30 @@
-export * from '../actual/middleware';
+export {
+  activityComponent,
+  ActivityPolymiddlewareProxy,
+  createActivityPolymiddleware,
+  useBuildRenderActivityCallback,
+  type ActivityPolymiddleware,
+  type ActivityPolymiddlewareHandler,
+  type ActivityPolymiddlewareHandlerResult,
+  type ActivityPolymiddlewareProps,
+  type ActivityPolymiddlewareProxyProps,
+  type ActivityPolymiddlewareRenderer,
+  type ActivityPolymiddlewareRequest,
+  type Polymiddleware
+} from 'botframework-webchat-api/middleware';
+
+export {
+  createErrorBoxPolymiddleware,
+  errorBoxComponent,
+  ErrorBoxPolymiddlewareProxy,
+  useBuildRenderErrorBoxCallback,
+  type ErrorBoxPolymiddleware,
+  type ErrorBoxPolymiddlewareHandler,
+  type ErrorBoxPolymiddlewareHandlerResult,
+  type ErrorBoxPolymiddlewareProps,
+  type ErrorBoxPolymiddlewareProxyProps,
+  type ErrorBoxPolymiddlewareRenderer,
+  type ErrorBoxPolymiddlewareRequest
+} from 'botframework-webchat-api/middleware';
+
+export { createActivityPolymiddlewareFromLegacy } from 'botframework-webchat-api/middleware';
