+++ packages/bundle/src/boot/iife/full-es5.ts
