+++ packages/bundle/src/boot/iife/minimal.ts
@@ -1,9 +1,9 @@
 import buildInfo from '../../buildInfo';
-import * as middleware from '../actual/middleware';
-import * as actual from '../actual/minimal.js';
+import * as middleware from '../exports/middleware';
+import * as minimal from '../exports/minimal.js';
 
 // Until we have a development-specific bundle, we are not shipping createStoreWithDevTools in bundle.
-const { createStoreWithDevTools: _createStoreWithDevTools, ...exports } = actual;
+const { createStoreWithDevTools: _createStoreWithDevTools, ...exports } = minimal;
 
 window['WebChat'] = Object.freeze({
   ...window['WebChat'],
