+++ packages/bundle/src/boot/iife/polyfill/modern.ts
