+++ packages/bundle/tsup.config.ts
@@ -55,9 +55,9 @@ export default defineConfig([
     },
     dts: false,
     entry: {
-      webchat: './src/boot/bundle/full.ts',
-      'webchat-es5': './src/boot/bundle/full-es5.ts',
-      'webchat-minimal': './src/boot/bundle/minimal.ts'
+      webchat: './src/boot/iife/full.ts',
+      'webchat-es5': './src/boot/iife/full-es5.ts',
+      'webchat-minimal': './src/boot/iife/minimal.ts'
     },
     esbuildPlugins: [...(commonConfig.esbuildPlugins ?? []), resolveReact],
     format: 'iife',
